add resolve_hostname(..) util function, refactor some utils into sub-modules
